+++ zhaquirks/linxura/button.py
@@ -6,6 +6,7 @@ from zigpy.profiles import zha
 from zigpy.quirks import CustomCluster, CustomDevice
 from zigpy.types import Addressing
 from zigpy.zcl import foundation
+from zigpy.zcl.foundation import ZCLHeader
 from zigpy.zcl.clusters.general import Basic
 import zigpy.zcl.clusters.security
 from zigpy.zcl.clusters.security import IasZone
@@ -35,12 +36,12 @@ from zhaquirks.const import (
 # from __init__ import LINXURA, LinxuraIASCluster
 from zhaquirks.linxura import LINXURA
 
-DOUBLE = 2
-HOLD = 3
-# LINXURA = "Linxura"
-SINGLE = 1
+# DOUBLE = 2
+# HOLD = 3
+# # LINXURA = "Linxura"
+# SINGLE = 1
 
-CLICK_TYPES = {SINGLE: "single", DOUBLE: "double", HOLD: "hold"}
+# CLICK_TYPES = {SINGLE: "single", DOUBLE: "double", HOLD: "hold"}
 
 BTN_1 = "Button 1"
 BTN_2 = "Button 2"
@@ -48,40 +49,40 @@ BTN_3 = "Button 3"
 BTN_4 = "Button 4"
 
 
-class LinxuraIASCluster(CustomCluster, zigpy.zcl.clusters.security.IasZone):
-    """Occupancy cluster."""
+# class LinxuraIASCluster(CustomCluster, zigpy.zcl.clusters.security.IasZone):
+#     """Occupancy cluster."""
 
-    def handle_cluster_request(
-        self,
-        hdr: foundation.ZCLHeader,
-        args: list[Any],
-        *,
-        dst_addressing: Optional[
-            Union[Addressing.Group, Addressing.IEEE, Addressing.NWK]
-        ] = None,
-    ):
-        """Handle a cluster command received on this cluster."""
-        # self.info(
-        #    "Linxura general request - handle_cluster_general_request: header: %s - args: [%s]",
-        #    hdr,
-        #    args,
-        # )
-        if hdr.command_id == 0:
-            # self.info(
-            #    "Linxura general request - state: %s",
-            #    args[0],
-            # )
-            state = args[0]
-            if state >= 4:
-                return
-            else:
-                event_args = {
-                    PRESS_TYPE: CLICK_TYPES[state],
-                    COMMAND_ID: hdr.command_id,
-                    ARGS: args,
-                }
-                action = f"button_{CLICK_TYPES[state]}"
-                self.listener_event(ZHA_SEND_EVENT, action, event_args)
+#     def handle_cluster_request(
+#         self,
+#         hdr: foundation.ZCLHeader,
+#         args: list[Any],
+#         *,
+#         dst_addressing: Optional[
+#             Union[Addressing.Group, Addressing.IEEE, Addressing.NWK]
+#         ] = None,
+#     ):
+#         """Handle a cluster command received on this cluster."""
+#         # self.info(
+#         #    "Linxura general request - handle_cluster_general_request: header: %s - args: [%s]",
+#         #    hdr,
+#         #    args,
+#         # )
+#         if hdr.command_id == 0:
+#             # self.info(
+#             #    "Linxura general request - state: %s",
+#             #    args[0],
+#             # )
+#             state = args[0]
+#             if state >= 4:
+#                 return
+#             else:
+#                 event_args = {
+#                     PRESS_TYPE: CLICK_TYPES[state],
+#                     COMMAND_ID: hdr.command_id,
+#                     ARGS: args,
+#                 }
+#                 action = f"button_{CLICK_TYPES[state]}"
+#                 self.listener_event(ZHA_SEND_EVENT, action, event_args)
 
 
 class LinxuraButton(CustomDevice):
